% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{ALE}
\alias{ALE}
\title{Generic function for Accumulated Local Effect (ALE)}
\usage{
ALE(o, ...)
}
\arguments{
\item{o}{the model we want to use to produce the ALE effect.}

\item{...}{arguments to be passed to methods.}
}
\description{
Generic function for producing ALE effects, to be plottied using the \code{plot} generic.
}
\references{
Apley, D.W., and Zhu, J, 2016. Visualizing the effects of predictor variables in black
box supervised learning models. arXiv preprint arXiv:1612.08468.
}
\seealso{
ALE.gam
}
