% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_family_cdf.R
\name{fix_family_cdf}
\alias{fix_family_cdf}
\title{Getting the CDF of a gam family}
\usage{
fix_family_cdf(fam)
}
\arguments{
\item{fam}{an object of class \code{family}.}
}
\description{
Some methods implemented in \code{mgcViz} require the c.d.f. of the response distribution.
This function takes a family object as input and returns the same object, but with the cdf
function added to the \code{$cdf} slot. Mainly for internal use.
}
