% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly.R
\name{ggplotly.plotSmooth}
\alias{ggplotly.plotSmooth}
\title{Convert to plotly}
\usage{
\method{ggplotly}{plotSmooth}(p, ...)
}
\arguments{
\item{x}{An object of class \code{plotSmooth}.}
}
\description{
Convert to plotly
}
