% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_ciBar.R
\name{l_ciBar}
\alias{l_ciBar}
\title{Adding confidence intervals to barplots}
\usage{
l_ciBar(level = 0.95, mul = NULL, ...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_errorbar}.}
}
\value{
an object of class \code{gamLayer}.
}
\description{
Adding confidence intervals to barplot.
}
