% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_ciLine.R
\name{l_ciLine}
\alias{l_ciLine}
\title{Add confidence curves to plot}
\usage{
l_ciLine(level = 0.95, mul = NULL, ...)
}
\arguments{
\item{level}{coverage level, should be in (0, 1).}

\item{mul}{number multiplied by the standard errors when calculating
standard error curves. By default \code{NULL}, if
set to a positive number it will over-ride \code{level}.}

\item{...}{graphical arguments to be passed to \code{ggplot2::geom_line}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
XXX
}
