% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_ciPoly.R
\name{l_ciPoly}
\alias{l_ciPoly}
\title{Add confidence band to smooth effect plot}
\usage{
l_ciPoly(level = 0.95, mul = NULL, ...)
}
\arguments{
\item{level}{coverage level, should be in (0, 1).}

\item{mul}{number multiplied by the standard errors when calculating
standard error surfaces. By default \code{NULL}, if
set to a positive number it will over-ride \code{level}.}

\item{...}{graphical arguments to be passed to \code{ggplot2::geom_polygon}.}
}
\value{
An object of class \code{gamLayer}
}
\description{
XXX
}
