% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_clusterLine.R
\name{l_clusterLine}
\alias{l_clusterLine}
\title{Clustered smooth effect plots}
\usage{
l_clusterLine(centers, cluFun = kmeans, a.clu = list(), ...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_line}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
Plotting clustered smooth effects.
}
