% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_coordContour.R
\name{l_coordContour}
\alias{l_coordContour}
\title{Add coordinate lines}
\usage{
l_coordContour(brLO = c(-8:9 * 20), brLA = c(-8:8 * 10), ...)
}
\arguments{
\item{...}{graphical arguments to be passed to  \code{ggplot2::geom_contour}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
XXX
}
