% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_dens.R
\name{l_dens}
\alias{l_dens}
\title{Add density of partial residuals to plot}
\usage{
l_dens(type, n = c(50, 50), bw = NULL, tol = 1e-06, trans = sqrt, ...)
}
\arguments{
\item{type}{if set to "cond" then the conditional residual density is plotted.
If set to "joint" the joint density of residuals and corresponding
covariate is plotted.}

\item{n}{vector of two positive integers, indicating the number of grid points
at which the density is evaluated on the x and y axes.}

\item{bw}{vector with two positive entried, indicating the bandwidth to be used
by the kernel density estimator along x and y.}

\item{tol}{small positive numerical tolerance. The estimated density at a certain
location is set to \code{NA} (hence it will appear white) when it falls
below \code{tol/sqrt(2*pi*sig)}, where \code{sig} is the standard
deviation of the residuals. Set \code{tol} to -1 plot the density on
the whole x-y plane, no matter how low it is.}

\item{trans}{the density on x-y is transformed using this function before being plotted.}

\item{...}{graphical arguments to be passed to \code{ggplot2::geom_raster}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
XXX
}
