% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_dens1D.R
\name{l_dens1D}
\alias{l_dens1D}
\title{Adding density estimate to a plot}
\usage{
l_dens1D(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_density}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
This layer adds a density estimate to a plot. It is mainly a
wrapper around \link[ggplot2:geom_density]{ggplot2::geom_density}.
}
\seealso{
See \link{check0D} for examples.
}
