% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitBar.R
\name{l_fitBar}
\alias{l_fitBar}
\title{Adding barplot to effect plots}
\usage{
l_fitBar(a.aes = list(), ...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_bar}.}
}
\value{
an object of class \code{gamLayer}.
}
\description{
Adding barplot to effect plots.
}
