% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitContour.R
\name{l_fitContour}
\alias{l_fitContour}
\title{Add fitted smooth effect contour}
\usage{
l_fitContour(...)
}
\arguments{
\item{...}{graphical arguments to be passed to  \code{ggplot2::geom_contour}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
XXX
}
