% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitLine.R
\name{l_fitLine}
\alias{l_fitLine}
\title{Add fitted smooth effect curve}
\usage{
l_fitLine(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_line}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
XXX
}
