% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitPoints.R
\name{l_fitPoints}
\alias{l_fitPoints}
\title{Adding fitted points to effect plots}
\usage{
l_fitPoints(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_point}.}
}
\value{
an object of class \code{gamLayer}.
}
\description{
Adding fitted points to smooth or parametric effect plots.
}
