% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitRaster.R
\name{l_fitRaster}
\alias{l_fitRaster}
\title{Add raster to smooth effect plot}
\usage{
l_fitRaster(pFun = function(.p) 1, noiseup = FALSE, mul = 1, ...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_raster}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
XXX
}
