% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_gridCheck2D.R
\name{l_gridCheck2D}
\alias{l_gridCheck2D}
\title{Checking GAM residuals along two covariates}
\usage{
l_gridCheck2D(gridFun = mean, bw = c(NA, NA), stand = TRUE,
  binFun = NULL, ...)
}
\arguments{
\item{gridFun}{function used to summarize the residuals in each bin. By default it is \code{mean}.}

\item{bw}{numeric vector giving bin width in both vertical and horizontal directions. See the \code{binwidth}
arguments in \code{?ggplot2::stat_summary_hex}. If left to \code{NA} is will set to one 20th
of the ranges of x and y.}

\item{stand}{if left to \code{TRUE} then the observed statistic in the i-th cell is normalized using
the simulated statistics in that same cell. That is we are actually plotting
\code{std_stat = (obs_stat-mean(sim_stat))/sd(sim_stat)}.}

\item{...}{graphical arguments to be passed to \code{ggplot2::stat_summary_hex}.}
}
\value{
An object of class \code{gamLayer}
}
\description{
XXX
}
\examples{
library(mgcViz);
set.seed(4124)
n <- 1e4
x <- rnorm(n); y <- rnorm(n);

# Residuals are heteroscedastic w.r.t. x
ob <- (x)^2 + (y)^2 + (1*abs(x) + 1)  * rnorm(n)
b <- bam(ob ~ s(x,k=30) + s(y, k=30), discrete = TRUE)
b <- getViz(b, nsim = 50)

# Don't see much by looking at mean
check2D(b, "x", "y") + l_gridCheck2D(gridFun = mean, bw = c(0.4, 0.4))

# Variance pattern along x-axis clearer now
check2D(b, "x", "y") + l_gridCheck2D(gridFun = sd, bw = c(0.4, 0.4))
 
}
