% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_hist.R
\name{l_hist}
\alias{l_hist}
\title{Adding histogram to a plot}
\usage{
l_hist(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_histogram}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
This layer adds a histogram to a plot. It is mainly a
wrapper around \link[ggplot2:geom_histogram]{ggplot2::geom_histogram}.
}
\seealso{
See \link{check0D} for examples.
}
