% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_points.R
\name{l_points}
\alias{l_points}
\title{Add residuals to smooth effect plot}
\usage{
l_points(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_point}.}
}
\value{
an object of class \code{gamLayer}.
}
\description{
XXX
}
