% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_poly.R
\name{l_poly}
\alias{l_poly}
\title{Add polygons to Markov random field plot}
\usage{
l_poly(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_polygon}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
Add polygons to Markov random field plot.
}
