% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_pvContour.R
\name{l_pvContour}
\alias{l_pvContour}
\title{Adding contour of p-values to 2D smooth effect plot}
\usage{
l_pvContour(pFun = identity, ...)
}
\arguments{
\item{...}{graphical arguments to be passed to  \code{ggplot2::geom_contour}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
Adding contour of p-values to 2D smooth effect plot
}
