% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_pvRaster.R
\name{l_pvRaster}
\alias{l_pvRaster}
\title{Adding heatmap of p-values to 2D smooth effect plot}
\usage{
l_pvRaster(pFun = identity, ...)
}
\arguments{
\item{pFun}{the transformations to be applied to the p-values before plotting.}

\item{...}{graphical arguments to be passed to \code{ggplot2::geom_raster}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
Adding heatmap of p-values to 2D smooth effect plot.
}
