% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_simLine.R
\name{l_simLine}
\alias{l_simLine}
\title{Add simulated smooth effect curves}
\usage{
l_simLine(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_line}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
This layer adds curves representing smooth effects simulated from the
posterior distribution.
}
\details{
This function uses transparency to avoid over-plotting.
This can be avoided by setting \code{alpha = 1} in the call to \code{l_simLine}.
}
\seealso{
See \link{plot.mgcv.smooth.1D} for examples.
}
