% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_vline.R
\name{l_vline}
\alias{l_vline}
\title{Adding vertical line to a plot}
\usage{
l_vline(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_vline}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
This layer adds a vertical to a plot. It is mainly a
wrapper around \link[ggplot2:geom_abline]{ggplot2::geom_vline}.
}
\seealso{
See \link{check0D} for examples.
}
