% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqgamV.R
\name{mqgamV}
\alias{mqgamV}
\title{Fit multiple QGAM models and get a mgamViz object}
\usage{
mqgamV(form, data, qu, lsig = NULL, err = NULL, aQgam = list(), aViz = list())
}
\arguments{
\item{form, data, qu, lsig, err}{same arguments as in \link[qgam:mqgam]{qgam::mqgam}.}

\item{aQgam}{list of further arguments to be passed to \link[qgam:mqgam]{qgam::mqgam}.}

\item{aViz}{list of arguments to be passed to \link{getViz}.}
}
\value{
An object of class "mgamViz" which can, for instance, be plotted using \link{plot.mgamViz}.
}
\description{
These are wrapper that fits multple QGAM models using \link[qgam:mqgam]{qgam::mqgam} and
converts it to a \code{mgamViz} object using the \link{getViz} function.
It is essentially a shortcut.
}
\examples{
library(mgcViz)
set.seed(2) ## simulate some data...
dat <- gamSim(2,n=500,dist="normal",scale=0.25)$data

# Fit GAM and get gamViz object
b <- mqgamV(y~s(x) + s(z) + I(x*z), data = dat, qu = c(0.25, 0.5, 0.75),
            aQgam = list(argGam = list(select = TRUE)), aViz = list("nsim" = 0))

# This is equivalent to doing
# 1. Fit QGAM
# b <- mqgam(y~s(x) + s(z) + I(x*z), data=dat, 
#            qu = c(0.25, 0.5, 0.75), argGam = list(select = TRUE))
# 2. Convert to gamViz object
# b <- getViz(b, nsim = 0)

# Either way, we all effects by doing
print(plot(b, allTerms = TRUE), pages = 1)


}
