% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgcv_smooth_2D.R
\name{plot.mgcv.smooth.2D}
\alias{plot.mgcv.smooth.2D}
\title{Plotting two dimensional smooth effects}
\usage{
\method{plot}{mgcv.smooth.2D}(x, n = 40, xlim = NULL, ylim = NULL,
  maxpo = 10000, too.far = 0.1, trans = identity, seWithMean = FALSE,
  unconditional = FALSE, ...)
}
\arguments{
\item{xlim}{if supplied then this pair of numbers are used as the x limits for the plot.}

\item{ylim}{if supplied then this pair of numbers are used as the y limits for the plot.}

\item{...}{currently unused.}
}
\description{
XXX
}
\examples{
library(mgcViz)
set.seed(2) ## simulate some data...
dat <- gamSim(1, n = 1000, dist = "normal", scale = 2)
b <- gam(y ~ s(x0) + s(x1, x2) + s(x3), data = dat, method = "REML")
b <- getViz(b)

# Plot 2D effect with noised-up raster, contour and rug for design points 
# Opacity is proportional to the significance of the effect
plot(sm(b, 2)) + l_fitRaster(pTrans = zto1(0.05, 2, 0.1), noiseup = TRUE) + 
  l_rug() + l_fitContour()  

# Plot contour of effect joint density of design points
plot(sm(b, 2)) + l_dens(type = "joint") + l_points() + l_fitContour() + 
  coord_cartesian(expand = FALSE) # Fill the plot
}
