% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ptermInteraction.R,
%   R/plot_multi_ptermInteraction.R
\name{plot.ptermInteraction}
\alias{plot.ptermInteraction}
\alias{plot.multi.ptermInteraction}
\title{Plotting parametric interactions}
\usage{
\method{plot}{ptermInteraction}(x, ...)

\method{plot}{multi.ptermInteraction}(x, ...)
}
\arguments{
\item{x}{a parametric interaction object, extracted using \link{pterm}.}

\item{...}{currently unused.}
}
\value{
Currently it returns \code{NULL}.
}
\description{
This function is here only to deal with parametric interactions (eg x0:fact), which
cannot be plotted at the moment.
}
