% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_si_smooth_1D.R
\name{plot.si.smooth.1D}
\alias{plot.si.smooth.1D}
\title{Plotting one dimensional single index effects}
\usage{
\method{plot}{si.smooth.1D}(x, n = 100, xlim = NULL, maxpo = 10000, trans = identity, ...)
}
\arguments{
\item{x}{a smooth effect object.}

\item{n}{number of grid points used to compute main effect and c.i. lines.
For a nice smooth plot this needs to be several times the estimated degrees of
freedom for the smooth.}

\item{xlim}{if supplied then this pair of numbers are used as the x limits for the plot.}

\item{maxpo}{maximum number of residuals points that will be used by layers such as
\code{resRug()} and \code{resPoints()}. If number of datapoints > \code{maxpo},
then a subsample of \code{maxpo} points will be taken.}

\item{trans}{monotonic function to apply to the smooth and residuals, before plotting.
Monotonicity is not checked.}

\item{...}{currently unused.}
}
\value{
An object of class \code{c("plotSmooth", "gg")}.
}
\description{
This method should be used to plot smooth effects
of class \code{"si.smooth.1D"}.
}
