% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plotDiff}
\alias{plotDiff}
\title{Generic plotting of differences}
\usage{
plotDiff(...)
}
\arguments{
\item{...}{arguments to be passed to methods. This first one will determine which
method will be called.}
}
\description{
Generic function for plotting differences between objects. Useful
mainly for plotting the differences between two smooth effects.
}
\seealso{
\link{plotDiff.mgcv.smooth.1D}, \link{plotDiff.mgcv.smooth.2D}, \link{plotDiff.sos.smooth}
}
