% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRGL_mgcv_smooth_MD.R
\name{plotRGL.mgcv.smooth.MD}
\alias{plotRGL.mgcv.smooth.MD}
\title{Visualizing 2D slice of smooth effects in 3D (interactively)}
\usage{
\method{plotRGL}{mgcv.smooth.MD}(x, fix, se = TRUE, n = 40,
  residuals = FALSE, type = "auto", maxpo = 1000, too.far = c(0, NA),
  xlab = NULL, ylab = NULL, main = NULL, xlim = NULL, ylim = NULL,
  se.mult = 1, shift = 0, trans = function(.x) {     .x },
  seWithMean = FALSE, unconditional = FALSE, ...)
}
\description{
XXX
}
\examples{
# Example 1: taken from ?mgcv::te, shows how tensor pruduct deals nicely with 
# badly scaled covariates (range of x 5\% of range of z )
library(mgcViz)
n <- 1e3
x <- rnorm(n); y <- rnorm(n); z <- rnorm(n)

ob <- (x-z)^2 + (y-z)^2 + rnorm(n)
b <- gam(ob ~ s(x, y, z))
v <- getViz(b)

plotRGL(sm(v, 1), fix = c("z" = 0))

rgl.close() # Close

plotRGL(sm(v, 1), fix = c("z" = 1), residuals = TRUE)

# We can still work on the plot, for instance change the aspect ratio
library(rgl)
aspect3d(1, 2, 1)

rgl.close() # Close

}
