% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRGL_nested2D.R
\name{plotRGL.nested2D}
\alias{plotRGL.nested2D}
\title{Visualizing 2D nested effects in 3D}
\usage{
\method{plotRGL}{nested2D}(
  x,
  se = TRUE,
  n = 40,
  residuals = FALSE,
  type = "auto",
  maxpo = 1000,
  too.far = 0,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  se.mult = 1,
  trans = identity,
  unconditional = FALSE,
  ...
)
}
\arguments{
\item{x}{a smooth effect object, extracted using \link{sm}.}

\item{se}{when TRUE (default) upper and lower surfaces are added to the plot at \code{se.mult}
(see below) standard deviations for the fitted surface.}

\item{n}{sqrt of the number of grid points used to compute the effect plot.}

\item{residuals}{if TRUE, then the (whole model) residuals will be added, as coloured spheres
on a plane below the fitted surface, at the observed \code{(si(x), exp(x))}
locations. See \link{plotRGL.mgcv.smooth.2D}.}

\item{type}{the type of residuals that should be plotted. See \link{residuals.gamViz}.}

\item{maxpo}{maximum number of residuals points that will be plotted.
If number of datapoints > \code{maxpo}, then a subsample of \code{maxpo} points will be taken.}

\item{too.far}{if greater than 0 then this is used to determine when a location is too far
from data to be plotted. The data are scaled into the unit square before deciding
what to exclude, and too.far is a distance within the unit square.
Unlike \link{plot.nested2D}, the default here is 0 (no exclusion), as in
\link{plotRGL.mgcv.smooth.2D}.}

\item{xlab}{if supplied then this will be used as the x label of the plot.}

\item{ylab}{if supplied then this will be used as the y label of the plot.}

\item{main}{used as title for the plot if supplied.}

\item{xlim}{if supplied then this pair of numbers are used as the limits of \code{si(x)}
(x axis) for the plot.}

\item{ylim}{if supplied then this pair of numbers are used as the limits of \code{exp(x)}
(y axis) for the plot.}

\item{se.mult}{a positive number which will be the multiplier of the standard errors
when calculating standard error surfaces.}

\item{trans}{currently unused (kept only for signature parity with
\link{plotRGL.mgcv.smooth.2D}, which accepts it but does not apply it either).}

\item{unconditional}{if \code{TRUE} then the smoothing parameter uncertainty corrected covariance
matrix is used to compute uncertainty bands, if available.
Otherwise the bands treat the smoothing parameters as fixed.}

\item{...}{currently unused.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
This method plots an interactive 3D representation of the whole effect
(all three parts added together, see \link{plot.nested2D}) of a two dimensional
nested smooth effect fitted with \code{gamFactory}. At the moment only
effects of type \code{"inter_le"}, that is \code{s(si(x), exp(x))}, are
supported. The output is built with the same internal renderer used by
\link{plotRGL.mgcv.smooth.2D}, so it has the same style: a red surface for the
fitted effect and, if \code{se = TRUE}, two blue wireframe surfaces at
\code{+/- se.mult} standard errors from it.
}
\examples{
\dontrun{
library(mgcViz)
library(gamFactory)
library(rgl)

# fit <- gam_nl(list(y ~ s_nest(X_l, X_e, trans = trans_inter_le()), ~ 1),
#               data = dat, family = fam_gaussian(), optimizer = "efs")
b <- getViz(fit)

plotRGL(sm(b, 1))

# Also show the (whole model) residuals
rgl.close()
plotRGL(sm(b, 1), residuals = TRUE)
}
}
\seealso{
\link{plotRGL.mgcv.smooth.2D}, \link{plot.nested2D}.
}
