% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plotRGL}
\alias{plotRGL}
\title{Generic RGL plotting function}
\usage{
plotRGL(x, ...)
}
\arguments{
\item{x}{the object we want to plot using the \code{rgl} package.}

\item{...}{arguments to be passed to methods.}
}
\description{
Generic function for producing an interactive RGL plot.
}
\seealso{
\link{plotRGL.mgcv.smooth.2D}, \link{plotRGL.mgcv.smooth.MD}
}
