% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_checkGam.R
\name{print.checkGam}
\alias{print.checkGam}
\title{Printing the output of check.gamViz}
\usage{
\method{print}{checkGam}(x, lay = NULL, ...)
}
\arguments{
\item{x}{the output of \code{check.gamViz}.}

\item{lay}{the \code{layout_matrix} passed to \link[gridExtra:grid.arrange]{gridExtra::grid.arrange}.}

\item{...}{further arguments to be passed to \code{grid.arrange}.}
}
\value{
Returns the output of \code{grid.arrange}, invisibly.
}
\description{
This method prints the output of \link{check.gamViz}.
}
