% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plotGam.R
\name{print.plotGam}
\alias{print.plotGam}
\title{Printing the output of plot.gamViz}
\usage{
\method{print}{plotGam}(x, ask = TRUE, pages = NULL, addLay = TRUE,
  ...)
}
\arguments{
\item{x}{an object of class \code{plotGam}.}

\item{ask}{should we ask before moving from one page to the next?}

\item{pages}{the number of pages over which to spread the output.}

\item{addLay}{if TRUE, and if the \code{$empty} slot of the \code{plotGam} object is TRUE,
we add some default layers to the plots, before printing. Does not have
any affect if the \code{plotGam} object already contains some layers.}

\item{...}{further arguments to be passed to \code{grid.arrange}.}
}
\value{
Returns the output of \link[gridExtra:grid.arrange]{gridExtra::grid.arrange}, invisibly.
}
\description{
This method prints the output of \link{plot.gamViz}.
}
