% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plotSmooth.R
\name{print.plotSmooth}
\alias{print.plotSmooth}
\title{Printing plots of smooth effects}
\usage{
\method{print}{plotSmooth}(x, ...)
}
\arguments{
\item{x}{an object of class \code{plotSmooth}.}

\item{...}{currently unused.}
}
\description{
Printing plots of smooth effects
}
