% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_qq_gam.R
\name{print.qqGam}
\alias{print.qqGam}
\title{Printing output of qq.gam}
\usage{
\method{print}{qqGam}(x, ...)
}
\arguments{
\item{x}{an object of class \code{qqGam}.}

\item{...}{currently unused.}
}
\description{
Printing output of qq.gam
}
