% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pterm.R
\name{pterm}
\alias{pterm}
\title{Extracting parametric effects from a GAM model}
\usage{
pterm(o, select)
}
\description{
This function can be used to extract a parametric effect from an object of
class \code{gamViz}.
}
