% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{qq}
\alias{qq}
\title{Generic QQ plots}
\usage{
qq(...)
}
\arguments{
\item{...}{arguments to be passed to methods. This first one will determine which
method will be called.}
}
\description{
Generic function for producing QQ-plots.
}
\seealso{
\link{qq.gamViz}
}
