% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals_gam.R
\name{residuals.gam}
\alias{residuals.gam}
\title{Generalized Additive Model residuals}
\usage{
\method{residuals}{gam}(object, type = "deviance", ...)
}
\description{
Wrapper around \code{mgcv::residuals.gam}.
}
