% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{shine}
\alias{shine}
\title{Generic shine function}
\usage{
shine(o, ...)
}
\arguments{
\item{o}{the object we want to transform into a \code{shiny} app.}

\item{...}{arguments to be passed to methods.}
}
\description{
Generic function for taking an object and transforming it into a \code{shiny} app.
}
\seealso{
shine.qqGam
}
