% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gam.R
\name{simulate.gam}
\alias{simulate.gam}
\title{Simulating responses from a GAM object}
\usage{
\method{simulate}{gam}(object, nsim = 1, seed = NULL, method = "auto",
  u = NULL, ...)
}
\arguments{
\item{nsim}{the number of simulated datasets. A positive integer.}

\item{u}{a matrix where each row is a vector of uniform random variables in (0, 1).}
}
\value{
a matrix where each row is a vector of simulated responses.
}
\description{
...
}
\examples{
library(mgcViz)

set.seed(2) ## simulate some data... 
dat <- gamSim(1,n=400,dist="normal",scale=2)
b <- gam(y~s(x0)+s(x1)+s(x2)+s(x3),data=dat)

# Simulate three vectors of responses
matplot(t(simulate(b, nsim = 3)), pch = 19, col = c(1, 3, 4)) 

}
