% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm}
\alias{sm}
\title{Extracting a smooth effect from a GAM model}
\usage{
sm(o, select)
}
\arguments{
\item{o}{an object of class \code{gamViz}, the output of a \link{getViz} call.}

\item{select}{index of the selected smooth or random effect.}
}
\value{
An object representing a smooth effect.
}
\description{
This function can be used to extract a smooth or random effect from an object of
class \code{gamViz}.
}
\seealso{
See \link{getViz} for examples.
}
