% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_qqGam.R
\name{zoom.qqGam}
\alias{zoom.qqGam}
\title{Allow to efficiently zoom on a qqGam object}
\usage{
\method{zoom}{qqGam}(o, xlim = NULL, ylim = NULL, discrete = NULL,
  ngr = 1000, adGrid = TRUE, CI = FALSE, worm = FALSE,
  show.reps = FALSE, a.qqpoi = list(), a.ablin = list(),
  a.cipoly = list(), a.replin = list(), ...)
}
\description{
XXX
}
\examples{
library(mgcViz);
set.seed(0)
n.samp <- 500
dat <- gamSim(1,n=n.samp,dist="binary",scale=.33)
p <- binomial()$linkinv(dat$f) ## binomial p
n <- sample(c(1,3),n.samp,replace=TRUE) ## binomial n
dat$y <- rbinom(n,n,p)
dat$n <- n
lr.fit <- bam(y/n ~ s(x0) + s(x1) + s(x2) + s(x3)
              , family = binomial, data = dat,
              weights = n, method = "REML")

set.seed(414)
o <- qq.gam(lr.fit, rep = 50, method = "simul1", CI = "normal")
o # This is the whole qqplot

# We can zoom in along x at little extra costs (most computation already done by qq.gam)
zoom(o, xlim = c(0, 1), show.reps = TRUE, 
     a.replin = list(alpha = 0.1), a.qqpoi =  list(shape = 19))
}
