% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zto1}
\alias{zto1}
\title{Function going from one to zero}
\usage{
zto1(o, a, m)
}
\description{
XXX
}
\examples{
library(mgcViz)
x <- seq(0, 1, by = 0.01)
plot(x, zto1(0.05, 2, 0.1)(x), ylim = c(0, 1))
lines(x, zto1(0.05, 1, 0.2)(x), col = 2)
lines(x, zto1(0.1, 3, 0)(x), col = 3)
}
