% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.gam.R
\name{check.gam}
\alias{check.gam}
\title{Some diagnostics for a fitted gam model}
\usage{
\method{check}{gam}(o, type = c("auto", "deviance", "pearson", "response",
  "tunif", "tnormal"), k.sample = 5000, k.rep = 200, maxpo = 10000,
  a.qq = list(), a.hist = list(), a.respoi = list(), a.reshex = list(),
  ...)
}
\arguments{
\item{o, }{A fitted \code{gam} object as produced by \code{\link[mgcv:gam]{mgcv::gam()}}.}

\item{type, }{Type of residuals, see \code{\link[mgcv:residuals.gam]{mgcv::residuals.gam()}}, used in all plots.}

\item{k.sample, }{Above this k testing uses a random sub-sample of data.}

\item{k.rep, }{How many re-shuffles to do to get p-value for k testing.}

\item{...}{Extra parameters.}

\item{rep, level, method, rl.col, rep.col, }{Arguments passed to \code{\link[=qq.gam]{qq.gam()}}.}
}
\value{
An object of class \code{check.gam}, which is simply a list of \code{ggplot} objects.
}
\description{
Takes a fitted gam object produced by \code{\link[mgcv:gam]{mgcv::gam()}} and produces some diagnostic
information about the fitting procedure and results. The default is to produce 4 residual plots,
some information about the convergence of the smoothness selection optimization, and
to run diagnostic tests of whether the basis dimension choises are adequate.
}
\note{
Help file is mainly from \link[mgcv:gam.check]{mgcv::gam.check} since this is a rewrite of \code{mgcv::gam.check}
function with ggplot2 library.
}
\examples{
library(ggplot2)
set.seed(0)
dat <- gamSim(1, n = 200)
b <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat)

# Checks using default options
check(b)

# Change some algorithmic and graphical parameters
check(b,
      a.qq = list(method = "tnorm", 
                  a.cipoly = list(fill = "light blue")), 
      a.respoi = list(size = 0.5), 
      a.hist = list(bins = 10))
}
