% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_bound.R
\name{l_bound}
\alias{l_bound}
\title{Add boundaries to smooth effect plot}
\usage{
l_bound(n = 200, ...)
}
\arguments{
\item{n}{number of discrete intervals along the boundary.}

\item{...}{graphical arguments to be passed to \code{ggplot2::geom_path}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
This layer adds boundaries to a smooth effect plot.
}
\seealso{
\link{plot.sos.smooth}
}
