% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_ciLine.R
\name{l_ciLine}
\alias{l_ciLine}
\title{Add confidence curves to plot}
\usage{
l_ciLine(mul = 2, ...)
}
\arguments{
\item{mul}{number multiplied by the standard errors when calculating
standard error curves or surfaces.}

\item{...}{graphical arguments to be passed to \code{ggplot2::geom_line}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
XXX
}
