% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_ciPoly.R
\name{l_ciPoly}
\alias{l_ciPoly}
\title{Add confidence band to smooth effect plot}
\usage{
l_ciPoly(mul = 2, ...)
}
\arguments{
\item{mul}{number multiplied by the standard errors when calculating
standard error curves or surfaces.}

\item{...}{graphical arguments to be passed to \code{ggplot2::geom_polygon}.}
}
\value{
An object of class \code{gamLayer}
}
\description{
XXX
}
