% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitBar.R
\name{l_fitBar}
\alias{l_fitBar}
\title{Adding barplot to effect plots}
\usage{
l_fitBar(a.aes = list(), ...)
}
\arguments{
\item{a.aes}{list of aesthetic mapping arguments that will be passed to
\code{ggplot2::geom_bar}. For instance we could set \code{a.aes=list("fill"="red")}
to change the colour of the barplot.}

\item{...}{graphical arguments to be passed to \link[ggplot2:geom_bar]{ggplot2::geom_bar}.}
}
\value{
an object of class \code{gamLayer}.
}
\description{
This layer adds a barplot to an effect plots. Mainly useful for factor or
binary effect plots.
}
\seealso{
See \link{plot.ptermFactor} for examples.
}
