% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitContour.R
\name{l_fitContour}
\alias{l_fitContour}
\title{Adding fitted effect contour lines}
\usage{
l_fitContour(...)
}
\arguments{
\item{...}{graphical arguments to be passed to  \code{ggplot2::geom_contour}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
This layer adds the contour lines corresponding to a fitted multidimensional effect.
}
\seealso{
See \link{plot.mgcv.smooth.2D}, \link{plot.mgcv.smooth.MD}, \link{plot.sos.smooth} and
\link{plotSlice} for examples.
}
