% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitLine.R
\name{l_fitLine}
\alias{l_fitLine}
\title{Add fitted smooth effect curve}
\usage{
l_fitLine(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_line}.}
}
\value{
An object of class \code{gamLayer}.
}
\description{
This layes add lines representing a single or a group of
parametric or smooth 1D effects.
}
\details{
When used in conjuction with \link{plot.fs.interaction.1D}, which plots smooth effects
of type \code{bs="fs"}, this function uses transparency to avoid over-plotting.
This can be avoided by setting \code{alpha = 1} in the call to \code{l_fitLine}.
}
\seealso{
See \link{plot.mgcv.smooth.1D}, \link{plot.ptermNumeric}, or
\link{plot.fs.interaction.1D} for examples.
}
