% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_fitPoints.R
\name{l_fitPoints}
\alias{l_fitPoints}
\title{Adding points representing the fitted effect}
\usage{
l_fitPoints(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_point}.}
}
\value{
an object of class \code{gamLayer}.
}
\description{
This function adds points representing the fitted effect. Mainly
useful for plotting factor effects.
}
\seealso{
See \link{plot.ptermFactor} for examples.
}
