% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_rug.R
\name{l_rug}
\alias{l_rug}
\title{Add rug to smooth effect plot}
\usage{
l_rug(...)
}
\arguments{
\item{...}{graphical arguments to be passed to \code{ggplot2::geom_rug}.}
}
\value{
an object of class \code{gamLayer}.
}
\description{
XXX
}
