% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiff_mgcv_smooth_2D.R
\name{plotDiff.mgcv.smooth.2D}
\alias{plotDiff.mgcv.smooth.2D}
\title{Plotting differences between 2D smooths}
\usage{
\method{plotDiff}{mgcv.smooth.2D}(s1, s2, n = 40, too.far = 0.1,
  trans = identity, unconditional = FALSE)
}
\description{
Plotting differences between 2D smooths.
}
\examples{
# Simulate data and add factors uncorrelated to the response
library(mgcViz)
set.seed(235)
dat <- gamSim(1,n=1500,dist="normal",scale=20)
dat$fac <- as.factor( sample(c("A1", "A2", "A3"), nrow(dat), replace = TRUE) ) 
dat$logi <- as.logical( sample(c(TRUE, FALSE), nrow(dat), replace = TRUE) ) 
bs <- "cr"; k <- 12
b <- gam(y ~ s(x2, x1, by = fac), data=dat)
o <- getViz(b, nsim = 0)

# Extract the smooths correspoding to "A1" and "A2" and plot their difference
pl <- plotDiff(s1 = sm(o, 1), s2 = sm(o, 2))
pl + l_fitRaster() + l_fitContour()

# Plot p-values for differences between the two smooths
pl + l_pvRaster() + l_pvContour()

}
