% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiff_sos_smooth.R
\name{plotDiff.sos.smooth}
\alias{plotDiff.sos.smooth}
\title{Plotting differences between smooths of sphere}
\usage{
\method{plotDiff}{sos.smooth}(s1, s2, n = 40, too.far = 0.1, phi = 30,
  theta = 30, scheme = 0, trans = identity, unconditional = FALSE)
}
\description{
Plotting differences between smooths of sphere.
}
