% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plotSmooth.R
\name{print.plotSmooth}
\alias{print.plotSmooth}
\title{Printing plots of smooth effects}
\usage{
\method{print}{plotSmooth}(x, addLay = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{plotSmooth}.}

\item{addLay}{if TRUE, and if the \code{$empty} slot of the \code{plotSmooth} object is TRUE
or NULL, we add some default layers to the plots, before printing. Does not have
any affect if the \code{plotSmooth} object already contains some layers (e.g. \code{l_rug()}).}

\item{...}{currently unused.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
This method prints objects of class \code{plotSmooth}.
}
