% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_qq_gam.R
\name{print.qqGam}
\alias{print.qqGam}
\title{Printing the output of qq.gamViz}
\usage{
\method{print}{qqGam}(x, ...)
}
\arguments{
\item{x}{an object of class \code{qqGam}.}

\item{...}{currently unused.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
This method prints the output of \link{qq.gamViz}.
}
